refactor: ディレクトリ名の変更をimport文へ反映
