+++ src/app/_containers/supplement_list/_components/createSupplementAction.ts
@@ -1,10 +1,8 @@
 "use server";
 
-import type { SupplementFormData } from "@/app/lib/types";
-import { supplementFormSchema } from "@/app/schema/supplementFormSchema";
+import { supplementFormSchema } from "@/app/_schema/supplementFormSchema";
 import { parseWithZod } from "@conform-to/zod";
 import { revalidateTag } from "next/cache";
-import { redirect } from "next/dist/server/api-utils";
 
 const BACKEND_API_URL = process.env.NEXT_PUBLIC_BACKEND_API_URL;
