+++ src/app/_containers/supplement_list/presentational.tsx
@@ -1,9 +1,9 @@
 "use client";
 
+import type { Supplement } from "@/app/_types/types";
 import { Skeleton } from "@nextui-org/skeleton";
 import { Pill } from "lucide-react";
 import React, { Suspense, useOptimistic, useTransition } from "react";
-import type { Supplement } from "../../../lib/types";
 import { CreateSupplementButton } from "./_components/CreateSupplementButton";
 import { EmptySupplementCard } from "./_components/EmptySupplementCard";
 import { deleteSupplementAction } from "./_components/deleteSupplementAction";
